refactor(tonk-schema): extract evaluate pipeline into `tonk_schema`
